Removes popper init
  * we do not rely on .popper.yml file anymore
  * misc. cleanup stuff
