+++ ci/test/actions-demo
@@ -1,23 +1,38 @@
 #!/bin/bash
 set -ex
-# shellcheck source=./ci/test/common
-source ./common
+## shellcheck source=./ci/test/common
+#source ./common
+#
+#delete_dir /tmp/gha-demo
+#git clone https://github.com/cplee/github-actions-demo /tmp/gha-demo
+#pushd /tmp/gha-demo
+#PHONY_SECRET=phony popper run --debug
+#popd
+#
+## test with an action defined on the root of the repo
+#delete_dir /tmp/npm
+#git clone https://github.com/actions/npm /tmp/npm
+#pushd /tmp/npm
+#DOCKER_USERNAME=foo DOCKER_PASSWORD=bar popper run "Integration Test"
+#popd
 
-delete_dir /tmp/gha-demo
-git clone https://github.com/cplee/github-actions-demo /tmp/gha-demo
-pushd /tmp/gha-demo
-PHONY_SECRET=phony popper run --debug
-popd
+# test with a remote action defined on the root of the repo
+init_test_repo
+cd $test_repo_path
+cat <<EOF > main.workflow
+workflow "action on root folder" {
+  resolves = "test"
+}
 
-# test with an action defined on the root of the repo
-delete_dir /tmp/npm
-git clone https://github.com/actions/npm /tmp/npm
-pushd /tmp/npm
-DOCKER_USERNAME=foo DOCKER_PASSWORD=bar popper run "Integration Test"
-popd
+action "test" {
+  uses = "poppperized/ansible@master"
+  args = "--version"
+}
+EOF
+
+ANSIBLE_SSH_KEY_DATA="dummydata" popper run
 
 # test with actions hosted on gitlab
-init_test_repo
 cd $test_repo_path
 cat <<EOF > main.workflow
 workflow "clone" {
