--- ci/test/init
@@ -1,10 +1,0 @@
-#!/bin/bash
-set -ex
-# shellcheck source=./ci/test/common
-source ./common
-init_test_repo
-cd "$test_repo_path"
-
-popper init
-test -f .popper.yml
-grep 'version: 2' .popper.yml
