+++ cli/popper/commands/cmd_ci.py
@@ -86,12 +86,18 @@ popper:
 )
 @pass_context
 def cli(ctx, service):
-    """Generates configuration files for distinct CI services.
+    """Generates configuration files for distinct CI services. This command
+    needs to be executed on the root of your Git repository folder.
     """
     if service not in ci_files:
         log.fail("Unrecognized service " + service)
 
-    project_root = scm.get_popper_root_folder()
+    project_root = scm.get_git_root_folder()
+
+    if project_root != os.getcwd():
+        log.fail(
+            'This command needs to be executed on the root of your'
+            'Git project folder.')
 
     for ci_file, ci_file_content in pu.get_items(ci_files[service]):
         ci_file_content = ci_file_content
